Adds TestLifecycleAware support in StartablePerSpecListener
