+++ modules/kotest/src/main/kotlin/org/testcontainers/containers/TestLifecycleAwareListener.kt
@@ -9,7 +9,7 @@ import org.testcontainers.lifecycle.TestLifecycleAware
 import java.net.URLEncoder
 import java.util.*
 
-class TestLifecycleAwareListener(startable: Startable) : TestListener {
+internal class TestLifecycleAwareListener(startable: Startable) : TestListener {
     private val testLifecycleAware = startable as? TestLifecycleAware
 
     override suspend fun beforeTest(testCase: TestCase) {
